Initial commit.
First baby steps. I'm trying to start by formulating the primitive
notions of Go (e.g. "board", "tile", "board state") as types. Knowing
myself, I expect this to be the hardest part and to get stuck on this
for quite a while.

At the moment I'm tring to figure out how I can describe these notions
with enough precision without committing to one particular
implementation, i.e. how to define what a board state is without
defining it to be a square matrix of entries of type `Color`.