Abstraction using interfaces.
It occurred to me that the abstraction I desire, namely to be able to
talk about the board state of a current game without commiting to
a certain representation like a square matrix, could be achieved by
using interfaces.

However, I didn't get very far because I have trouble using interfaces
with parameters.