Points on boards and Ajacency.
I've introduced a type `Point` in order to describe the points that lie
on a board, and I've introduced a type `Adjacent a b` that describes the
ways in which points `a, b : Point board` can be adjacent.

I've also decided that there is probably no harm in sticking to
a specific implementation of a board game state type (or a board point
type for that matter), as long as board states are constructed in an
abstract way that does not make direct reference to row/col coordinates,
but instead relies on the primitive notion of a point.

Finally, I reasoned that the basic primitives that are common to all
versions of Go should be strictly separated from the individual rulesets
(Japanese, Chinese, logical, ...) themselves.