Logical pseudo-rules.
Written down a modified version of the logical go rules by J. Tromp in
order to make progress with the formalization of the rules.

As there are written now, there are almost immediately ripe for turning
them into code. Some issues remain however:

* The definition of what constitutes a valid move, more specifically
rules 8.e), is somewhat imprecise. Maybe it could be formalized as
follows

    S2.coloring = clear(col(S1.Player), clear(col(opp(S1.Player)), color(col(S1.Player), P, S1.coloring)))

  Then one would need to introduce the function `clear` and `color` and
  fix their semantics.
* That Go is played on a grid of `n`x`n` points is completely immaterial
  to the rules. The only thing that is need is a *finite* set of points
  together with an adjacency relation on it.