Streamlined rules.
I've streamlined the rules a bit. The most significant change is that
I've switched to a more abstract definition of what a board is. The
reason for that is not that I think that three-dimensional go or any of
its more weird variants are that interesting but that the rules are more
natural to formulate in this generality. To wit, the fact that the board
and its relation of adjacency are that of a square grid of points
together with the adjacency induced by the taxi metric are completely
immaterial to the rules.