Towards a new abstract board type.
I've started reworking the definitions towards the more abstract (and
imo better-in-the-long-run) definition of "board" as given in the recent
version of `Rules.md`.

The definition of `Board'`---which is the new version of `Board` (I've
yet to remove the old one)---given at the moment might be subject to
change as I'm not sure whether I like it or not.