Simpler approach to relations.
It occurred to me that while viewing a relation as a function of type

    ty -> ty -> Type

seems somewhat natural, this also complicates things unnecessarily.
Therefore I've decided that it's probably better to simply view them as
functions

    ty -> ty -> Bool

At the moment, the old code is only uncommented for I wanted to have
a record of it in case I change my mind. This code will be removed in
the next commit.